% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.R
\name{callFacebookAPI}
\alias{callFacebookAPI}
\title{Facebook APIを呼び出す}
\usage{
callFacebookAPI(path, token = NULL, api_version = "v3.2")
}
\arguments{
\item{path}{呼び出したいAPIエンドポイントのパスからバージョン部分を抜いたもの。例: "/me?fields=id,name"}

\item{token}{アクセストークンをTokenオブジェクトまたは文字列で渡す}

\item{api_version}{APIバージョン。例: "v3.2"}
}
\value{
APIレスポンスのJSONをリストオブジェクトとしてパースしたもの
}
\description{
Facebook APIを呼び出す
}
\examples{
callFacebookAPI("/me?fields=id,name")
}
