% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook_Token2.0_factory.R
\docType{data}
\name{facebook_Token2.0_factory}
\alias{facebook_Token2.0_factory}
\title{Token2.0オブジェクトを作成するFactoryクラス}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
token_factory <- facebook_Token2.0_factory$new(consumer_key = "XXXXXX", consumer_secret = "XXXXXXXXXX")
}
\description{
Token2.0オブジェクトを作成するFactoryクラス
}
\examples{
# Facebookアプリを作成し、リダイレクトURIに https://127.0.0.1:1410/ を加えておく。HTTPSであることに注意。
token_factory <- facebook_Token2.0_factory$new(consumer_key = "0000000000", consumer_secret = "00000000000000000000", redirect_uri = "https://127.0.0.1:1410/")

# 認証ページURLを作成する
auth_url <- token_factory$create_authorize_url()

auth_url
# 認証ページURL auth_url をブラウザで開いて認証を行い、リダイレクトURLをブラウザのURLバーから読み取る

# 読み取ったリダイレクトURLを格納する
redirect_url <- 'https://127.0.0.1:1410/?code=XXXXXXXXXXXXX&state=AAAAAAA#_=_'

# リダイレクトURLをパースする
params <- token_factory$parse_redirect_url(redirect_url)
# アクセストークンを取得する
facebook_token <- token_factory$create_token(params)

}
\keyword{datasets}
