% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{data}
\name{FacebookAuthentication}
\alias{FacebookAuthentication}
\title{R6 environment to store authentication credentials}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FacebookAuthentication
}
\description{
Used to keep persistent state.
}
\keyword{datasets}
