% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{getLoadedAccessToken}
\alias{getLoadedAccessToken}
\title{メモリにロードされているアクセストークンを取得する}
\usage{
getLoadedAccessToken(name)
}
\arguments{
\item{name}{文字列. "google", "facebook" または "twitter"}
}
\value{
Tokenオブジェクト.
}
\description{
メモリにロードされているアクセストークンを取得する
}
