% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{getLoadedAccessTokens}
\alias{getLoadedAccessTokens}
\title{メモリにロードされているアクセストークンを取得する}
\usage{
getLoadedAccessTokens()
}
\value{
Tokenオブジェクトのリスト。 .$google, .$facebook, .$twitter
}
\description{
メモリにロードされているアクセストークンを取得する
}
\examples{
token_list <- getLoadedAccessTokens()
print(token_list$google)

}
