% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gftShinyAuthR-package.R
\docType{package}
\name{gftShinyAuthR-package}
\alias{gftShinyAuthR}
\alias{gftShinyAuthR-package}
\title{gftShinyAuthR: Google, Facebook and Twitter Authentication on Shiny Apps}
\description{
Google, Facebook and Twitter Authentication on Shiny Apps
}
\author{
\strong{Maintainer}: Masaki Terashi \email{terashi@addix.co.jp}

}
\keyword{internal}
