% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter_Token1.0_factory.R
\docType{data}
\name{twitter_Token1.0_factory}
\alias{twitter_Token1.0_factory}
\title{Token1.0オブジェクトを作成するFactoryクラス}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
token_factory <- twitter_Token1.0_factory$new(consumer_key = "XXXXXX", consumer_secret = "XXXXXXXXXX")
}
\description{
Token1.0オブジェクトを作成するFactoryクラス
}
\examples{
# インスタンス化する
token_factory <- twitter_Token1.0_factory$new(consumer_key = "XXXXXX", consumer_secret = "XXXXXXXXXX")
# 認証ページURLを作成する
auth_url <- token_factory$create_authorize_url()
# 認証ページURL auth_url をブラウザで開いて認証を行い、リダイレクトURLをブラウザのURLバーから読み取る
# 読み取ったリダイレクトURLを格納する
redirect_url <- 'http://127.0.0.1:1410/?oauth_token=fQvZDgAAAAAAXXXXX&oauth_verifier=XXXXXXXXXX'
# リダイレクトURLをパースする
params <- token_factory$parse_redirect_url(redirect_url)
# アクセストークンを取得する
twitter_token <- token_factory$create_token(params)

}
\keyword{datasets}
